requirements.txt
django
djangorestframework
django-filter
psycopg2
